feat: use HTMX modals for checkin/checkout
