+++ door_tracker/midas/api.py
@@ -318,125 +318,6 @@ def healthcheck(request):
     return Response(s.data)
 
 
-@dataclass
-class CheckoutResponse:
-    date: datetime
-
-
-class CheckoutRequestSerializer(Serializer):
-    time = DateTimeField()
-
-    def create(self, validated_data):
-        return validated_data['time']
-
-
-class CheckoutResponseSerializer(Serializer):
-    status = HiddenField(default='ok')
-    date = DateTimeField()
-
-    def create(self, validated_data):
-        return CheckoutResponse(**validated_data)
-
-
-@extend_schema(
-    operation_id='checkout',
-    request=CheckoutRequestSerializer,
-    responses={
-        201: CheckoutResponseSerializer,
-        400: APIResponseSerializer,
-        404: APIResponseSerializer,
-    },
-)
-@api_view(['POST'])
-def checkout(request):
-    """Check out remotely
-
-    This endpoint is called from frontend to check a user out remotely.
-    """
-    serializer = CheckoutRequestSerializer(data=request.data)
-    if not serializer.is_valid():
-        return serializer_error(serializer)
-    checkout_time = serializer.save()
-
-    if checkout_time > timezone.now():
-        res = APIResponse.error('Cannot checkout in the future')
-        s = APIResponseSerializer(res)
-        return Response(s.data, status=400)
-
-    try:
-        with transaction.atomic():
-            session = Session.objects.get(
-                user=request.user,
-                checkin__time__date=checkout_time,
-                checkout__isnull=True,
-            )
-            checkout = Checkout.objects.create(
-                type=LogType.REMOTE,
-                time=checkout_time,
-                session=session,
-            )
-    except Session.DoesNotExist:
-        res = APIResponse.error('There are no matching sessions')
-        s = APIResponseSerializer(res)
-        return Response(s.data, status=404)
-    else:
-        res = CheckoutResponse(date=checkout.time)
-        s = CheckoutResponseSerializer(res)
-        return Response(s.data, status=201)
-
-
-@dataclass
-class CheckinResponse:
-    date: datetime
-
-
-class CheckinResponseSerializer(Serializer):
-    status = HiddenField(default='ok')
-    date = DateTimeField()
-
-    def create(self, validated_data):
-        return CheckinResponse(**validated_data)
-
-
-@extend_schema(
-    operation_id='checkin',
-    responses={
-        201: CheckinResponseSerializer,
-        400: APIResponseSerializer,
-        403: APIResponseSerializer,
-    },
-)
-@api_view(['POST'])
-def checkin(request):
-    """Check in remotely
-
-    This endpoint is called from frontend to check a user in remotely.
-    """
-    checkin_time = timezone.now()
-
-    if Session.objects.filter(user=request.user, checkout__isnull=True).exists():
-        res = APIResponse.error('User already checked in')
-        s = APIResponseSerializer(res)
-        return Response(s.data, status=403)
-
-    try:
-        with transaction.atomic():
-            new_session = Session.objects.create(user=request.user)
-            Checkin.objects.create(
-                type=LogType.REMOTE,
-                time=checkin_time,
-                session=new_session,
-            )
-    except Exception as e:
-        res = APIResponse.error(f'Failed to check in: {str(e)}')
-        s = APIResponseSerializer(res)
-        return Response(s.data, status=400)
-
-    res = CheckinResponse(date=checkin_time)
-    s = CheckinResponseSerializer(res)
-    return Response(s.data, status=201)
-
-
 class ExportSessionsResponseSerializer(Serializer):
     pass
