+++ door_tracker/midas/templates/midas/base.html
@@ -1,4 +1,4 @@
-{% load static %}
+{% load static modal %}
 
 <!doctype html>
 <html lang="en">
@@ -50,8 +50,8 @@
 
         <h3 class="sidebar__section">ACTIONS</h3>
         <a class="sidebar__button" href="{% url 'midas:export_user' %}">Export Sessions</a>
-        <a class="sidebar__button" href="#" onclick="openCheckinModal()">Remote Checkin</a>
-        <a class="sidebar__button" href="#" onclick="openCheckoutModal()">Remote Checkout</a>
+        <a class="sidebar__button" href="{% querystring modal='checkin' %}" hx-boost="true">Remote Checkin</a>
+        <a class="sidebar__button" href="{% querystring modal='checkout' %}" hx-boost="true">Remote Checkout</a>
         <a class="sidebar__button" href="{% url 'midas:logout' %}">Log Out</a>
       </nav>
 
@@ -65,149 +65,27 @@
 
     </div>
 
-    <div id="checkoutModal" class="modal" style="display:none">
-      <form id="checkoutForm" class="modal__content">
-        <h2 class="modal__title">Remote Checkout</h2>
-        <label for="checkoutTime">Select Checkout Time:</label>
-        <input
-          type="datetime-local"
-          id="checkoutTime"
-          name="checkout_time"
-          required
-          step=1
-          max="{% now 'c' %}"
-          style="width: 100%;"
-        />
-        <div class="modal__actions" style="margin-top: 1rem;">
-          <button type="submit" class="modal__action modal--primary">Checkout</button>
-          <button type="button" class="modal__action modal--secondary" onclick="closeCheckoutModal()">Cancel</button>
-        </div>
-      </form>
-    </div>
-
-
-
-    <div id="checkinModal" class="modal" style="display:none">
-      <div class="modal__content">
-        <h2 class="modal__title">Remote Check-in</h2>
-        <p>Do you want to remotely check in with the current time?</p>
-        <div class="modal__actions">
-          <button type="button" class="modal__action modal--primary" onclick="confirmCheckin()">Yes</button>
-          <button type="button" class="modal__action modal--secondary" onclick="closeCheckinModal()">No</button>
-        </div>
+    {% modal 'checkin' 'midas:checkin' 'Remote Check-in' %}
+      <p>Do you want to remotely check in with current time?</p>
+      <div class="modal__actions">
+        <button class="modal__action modal--primary">Checkin</button>
+        <a class="modal__action modal--secondary" href="{% querystring modal=None %}" hx-boost="true">Cancel</a>
       </div>
-    </div>
+    {% endmodal %}
 
+    {% modal 'checkout' 'midas:checkout' 'Remote Checkout' %}
+      <label for="checkoutTime">Select Checkout Time:</label>
+      <input type="datetime-local" id="checkoutTime" name="time" required step=1 value="{% now 'Y-m-d\TH:i' %}" max="{% now 'Y-m-d\TH:i' %}" style="width: 100%;" />
+      <div class="modal__actions" style="margin-top: 1rem;">
+        <button type="submit" class="modal__action modal--primary">Checkout</button>
+        <a class="modal__action modal--secondary" href="{% querystring modal=None %}" hx-boost="true">Cancel</a>
+      </div>
+    {% endmodal %}
 
     <div class="messages">
       {% for message in messages %}
         <p{% if message.tags %} class="messages__item messages--{{ message.tags }}"{% endif %}>{{ message }}</p>
       {% endfor %}
     </div>
-
-    <script>
-      function showMessage(type, text) {
-        const message = document.createElement('p')
-        message.classList.add('messages__item', `messages--${type}`)
-        message.innerText = text
-        document.querySelector('.messages').appendChild(message)
-      }
-
-      function openCheckoutModal() {
-        checkoutModal.style.display = 'flex'
-      }
-
-      function closeCheckoutModal() {
-        checkoutModal.style.display = 'none'
-        checkoutForm.reset()
-      }
-
-      checkoutForm.addEventListener('submit', async (e) => {
-        e.preventDefault()
-
-        try {
-          const response = await fetch("{% url 'midas:checkout' %}", {
-            method: 'POST',
-            headers: {
-              'Content-Type': 'application/json',
-              'X-CSRFToken': '{{ csrf_token }}',
-            },
-            body: JSON.stringify({
-              time: new Date(checkoutTime.value).toISOString(),
-            }),
-          })
-
-          const data = await response.json()
-
-          if (response.ok) {
-            showMessage('success', `Successfully checked out at ${data.date}`)
-            setTimeout(() => {
-              window.location.reload()
-            }, 2000)
-          } else {
-            showMessage('error', data.message || 'An error occurred')
-          }
-        } catch (error) {
-          showMessage('error', 'Network error: ' + error.message)
-        }
-      })
-    </script>
-    <script>
-      const checkinModal = document.getElementById('checkinModal')
-
-      function openCheckinModal() {
-        checkinModal.style.display = 'flex'
-      }
-
-      function closeCheckinModal() {
-        checkinModal.style.display = 'none'
-      }
-
-      async function confirmCheckin() {
-        closeCheckinModal()
-
-        try {
-          const response = await fetch("{% url 'midas:checkin' %}", {
-            method: 'POST',
-            headers: {
-              'Content-Type': 'application/json',
-              'X-CSRFToken': '{{ csrf_token }}',
-            },
-            body: JSON.stringify({
-              time: new Date().toISOString(),
-            }),
-          })
-
-          const data = await response.json()
-
-          if (response.ok) {
-            showMessage('success', `Successfully checked in at ${data.date}`)
-            setTimeout(() => {
-              window.location.reload()
-            }, 2000)
-          } else {
-            showMessage('error', data.message || 'An error occurred during check-in')
-          }
-        } catch (error) {
-          showMessage('error', 'Network error: ' + error.message)
-        }
-      }
-    </script>
-    <script>
-      function openCheckoutModal() {
-        const now = new Date()
-        const offset = now.getTimezoneOffset()
-        const localISOTime = new Date(now.getTime() - offset * 60000).toISOString().slice(0, 19)
-        document.getElementById('checkoutTime').value = localISOTime
-
-        checkoutModal.style.display = 'flex'
-      }
-
-      function closeCheckoutModal() {
-        checkoutModal.style.display = 'none'
-        checkoutForm.reset()
-      }
-    </script>
-
   </body>
 </html>
